+++ app/src/main/java/com/android/artic/repository/Module.kt
@@ -3,6 +3,6 @@ package com.android.artic.repository
 import org.koin.dsl.module
 
 val module = module {
-    // singleton 으로 앱 전체에서 동일한 repository 에 접근할 수 있도록 제작
+    // design singleton for same accessibility in every where in this app
     single { ArticRepository() }
 }
