+++ app/src/main/java/com/android/artic/ui/home/HomeActivity.kt
@@ -26,6 +26,7 @@ class HomeActivity : BaseActivity() {
                 adapter.data = it
                 adapter.notifyDataSetChanged()
             }.apply { addDisposable(this) }
+
     }
 }
