+++ app/src/main/java/com/android/artic/ui/home/data/HomeCardKind.kt
@@ -1,0 +1,5 @@
+package com.android.artic.ui.home.data
+
+enum class HomeCardKind {
+    ARCHIVE, LINK
+}
